requirements.txt
numpy==1.24.1
opencv_python==4.8.0.76
Pillow==10.0.0
scikit_learn==1.3.0
scipy==1.11.1
torch==2.0.1+cu118
torchattacks==3.4.0
torchvision==0.15.2+cu118
umap==0.1.1